Updated the helpers directory of lte module

+++ src/lte/helper/lte-helper.h
@@ -694,8 +694,8 @@ class LteHelper : public Object
      */
     void DoComponentCarrierConfigure(uint32_t ulEarfcn,
                                      uint32_t dlEarfcn,
-                                     uint16_t ulbw,
-                                     uint16_t dlbw);
+                                     uint8_t ulbw,
+                                     uint8_t dlbw);
     /**
      * Create an eNodeB device (LteEnbNetDevice) on the given node.
      * \param n the node where the device is to be installed
